Add python function
Provides a python 2.7 function to write the authentication json string.
It uses python's `subprocess` module to call the sign_secp224k1
executable and passes the private key and message SHA224 hashes to the
standard in.
An example is provided that uses the values given at:
https://github.com/coinfloor/API/blob/master/AUTH.md
To run it, make sure the sign_secp224k1 executable is in the same folder
as the authenticate.py file, and run  `python authenticate.py`.
It uses only standard python modules

I tested this function in a live environment, and I could successfully
authenticate with my details.